util/TimeParser: wrapper for strptime()
Move code from SongFilter.cxx.
